requirements.txt
flask-sqlacodegen

